+++ client/client.go
@@ -84,7 +84,6 @@ type Client interface {
 	Delete(ctx context.Context, tableName string, rowKey []*table.Column, operationOpts ...OperationOption) (int64, error)
 	// Get a record by rowKey.
 	Get(ctx context.Context, tableName string, rowKey []*table.Column, getColumns []string, operationOpts ...OperationOption) (SingleResult, error)
-	Get(ctx context.Context, tableName string, rowKey []*table.Column, getColumns []string, opts ...ObkvOption) (SingleResult, error)
 	// Query records by rangePairs.
 	Query(ctx context.Context, tableName string, rangePairs []*table.RangePair, opts ...ObkvQueryOption) (QueryResultIterator, error)
 	// NewBatchExecutor create a batch executor.
