+++ protocol/query.go
@@ -20,6 +20,9 @@ package protocol
 import (
 	"bytes"
 
+	"github.com/pkg/errors"
+
+	"github.com/oceanbase/obkv-table-client-go/table"
 	"github.com/oceanbase/obkv-table-client-go/util"
 )
 
@@ -86,7 +89,7 @@ func NewObTableQueryWithParams(batchSize int32) *ObTableQuery {
 }
 
 func NewObTableQueryWithKeyRanges(startKeyColumns []*table.Column, endKeyColumns []*table.Column) (*ObTableQuery, error) {
-	obNewRange, err := NewObNewRangeWithParams(startKeyColumns, endKeyColumns)
+	obNewRange, err := NewObNewRangeWithColumns(startKeyColumns, endKeyColumns)
 	if err != nil {
 		return nil, errors.WithMessage(err, "create ob new range")
 	}
