+++ protocol/query_and_mutate_test.go
@@ -37,7 +37,7 @@ func TestObTableQueryAndMutateEncodeDecode(t *testing.T) {
 	obNewRanges := make([]*ObNewRange, 0, randomLen)
 	for i := 0; i < randomLen; i++ {
 		obNewRange := NewObNewRange()
-		obNewRange.SetTableId(int64(rand.Uint64()))
+		obNewRange.SetTableId(rand.Uint64())
 		obNewRange.SetBorderFlag(ObBorderFlag(rand.Intn(255)))
 		randomLen = rand.Intn(100)
 		startKey := make([]*ObObject, 0, randomLen)
