+++ protocol/query_test.go
@@ -36,7 +36,7 @@ func TestObTableQueryEncodeDecode(t *testing.T) {
 	obNewRanges := make([]*ObNewRange, 0, randomLen)
 	for i := 0; i < randomLen; i++ {
 		obNewRange := NewObNewRange()
-		obNewRange.SetTableId(int64(rand.Uint64()))
+		obNewRange.SetTableId(rand.Uint64())
 		obNewRange.SetBorderFlag(ObBorderFlag(rand.Intn(255)))
 		randomLen = rand.Intn(100)
 		startKey := make([]*ObObject, 0, randomLen)
